Moved ParserConfigConstants enum from common Constants.java to RegularExpressionsParser Class
